Adding a set of functions for a new method to find the generic...
The main function of those newly added is: cseq.
This function only take a sequence an automatically extract a pattern.
The method extract pattern from the dictionnary of patterns build with
the LZW algorithm...

This method seem to be much more powerfull when testing on random data
with the mseq.py file. :+1:
TODO: continue to benchmark...